Rework assembler format chunk hierarchy
This commit changes the way assembler format chunks are used. Instead of having
a single sequence of chunks, we keep a tree of chunks that reflects somewhat common
structure of assembly language(s) - that is, instruction is notated as a mnemonic
followed by zero or more operands separated by comma.

The root of the tree is always `AcAsmFormat`.

Also, the parsing of assembly has been relaxed a bit for allowing extra (or no) spaces
as hinted in ArchC manual (which is arguably out-dated).