Refactor disassembly of builtins
This commit splits disassembling of `%imm` and `%exp` chunks into two
parts:

 1. decoding the value (which includes both extracting value from
    bitfields and possibly inverse-transforming it according to modifier)

 2. printing the decoded value

The reason for this is that modifier defines how to transform the value
and `%imm` and `%exp` defines whether the value should be printed as
unsigned (hex) or signed (decimal).

As of now, this is only tested with RISC-V, more thinking / testing is
needed to make sure it works for all other ISAs too.

Do not merge as is.