Unify `#binaryEncoding` on format, instruction declaration and instruction itself
so that all these objects 'know' the binary encoding, which is either
fully symbolic (for instruction format), partially symbolic (for instruction
declaration or instantiated instruction) or fully concrete (for fully
grounded instruction). This also makes the API more consistent.

As a side effect, this allows for "incremental grounding" of instructions
in various stages of the compiler - see (new) test `#testInEnvironment`.
This is actually used in Tinyrossa.