Refactor `ProcessorInstruction >> #disassemble`
This commit changes the way `ProcessorInstruction >> #disassemble` picks
the "most specific" syntax form for given (concrete) instruction.

The original code picked the format with most "fixed" operands - this
mostly worked, but not always. Consider following formats:

    tw.set_asm("tw %imm, %reg, %reg", to, ra, rb);
    tw.set_asm("tw%[trapcond] %reg, %reg", to, ra, rb);

Clearly, `tweq 0,1` is preferable over `tw 4, 0, 1` but both bind 3
operands and have no "fixed" operands.

This commit changes the selection so that it selects the one with least
operands at assembly syntax level and if two formats have same number
of operands, prefer the one with more "fixed" operands over the other
(See `#isMoreOrEquallySpecificThan:`).

In the above case it'd pick the former since it has 2 operands and the 
latter has 3 operands.

Now, the specificity of syntax formats is a static property, so if we 
keep formats sorted by number of arguments, we can just try to
disassemble once by one and return the first that does not fail will be
the one with least number of operands. This commit changes the code to
do exactly this.