Refactor instruction decoding
This commit refactors the way instructions are decoded in
`AcProcessorDescription >> decodeBits:`:

  (i) Defer complete logic to instruction decoder object.
 (ii) Create new (base) class `AcInstructionDecoder` which should
      work with all ISA (models), even those with no single opcode field.
(iii) Add a custom subclass - `AcOpcodeDecoder` - that speeds up decoding
      for ISAs with single opcode field by limiting search to only
      instruction declarations with that opcode.

Of course, the base `AcInstructionDecoder` is slow because it tries to
match given encoding against every single instruction - but it is general
and should always work.

In the future we may introduce generalized multi-stage decoder that
configures itself from PDL without relying on existence of opcode field
but that's left as future work.