Use correct sort for x in #testPowerAddiSymbolic
In this line:

self assert: (instr fieldValue: 'd') equals: 'x'

the "instr fieldValue: 'd'" is a Z3Int, so the #= down in #assert:equals:
coerces the 'x' to a Z3Int.  So now we have two 'x' symbols: a BV16 'x'
and an Int 'x', which are not necessarily equal.
The correct thing to do is to start from the same BV 'x',
and compare the Int interpretation of `d` to the signed Int
interpretation of 'x'.