Fix failing `#testTrapiSubstitution`
Work on PLE in MA required *NOT* to `simplify` ASTs. This in turn
caused `#testTrapiSubstitution` to break because it assumed the bitvector
is simplified in a certain way.

To fix this, this commit rewrites the test in a way that does not depend
on particular way the bitvector get simplified.