Update makefiles and README after merging in PDLs
As PDLs are now part of Pharo-ArchC repository, there's no need to check
out PDL. This commit updates makefiles and README accordingly.

There's one issue with RISC-V PDL as it uses to upsteam `riscv-opcodes`
to provide list of instructions and encodings. Technically it is an
submodule so a special care is needed to make sure it is checked out.
This is not a problem when one is using makefiles as they care for this,
but it is a problem when relying solely on Iceberg / Metacello to load
ArchC since there's no support for submodules.