Fix 2 tests in `InstructionFormatTest` on Smalltalk/X
This commit tests `InstructionFormatTest >> #testMostVariedOperandsARM`
and `#testMostVariedOperandsSPARC` for Smalltalk/X - the problem was
that `Dictionary >> values` gives an `Array` on Pharo and OC on St/X.

The reason this worked so far was that Smalltalk/X had a hack in
`assert:equals:` to treat comparison of `Array` and `OrderedCollection`
specially. While convenient, this is non-standard behavior and hinders
portability. Therefore it will be removed from Smalltalk/X and this
commit proactively fixes the test.